Add a comment about lifecycle management.
If some processing function decides to free the event structs for
something that has just completed, we'll end up with a kevent pointing
to something that no longer exists.

So either we need to keep them around until the end of the kqueue()
call, where we can just cancel them, or we have to walk the
kqueue completed list and mark them as canceled.

I'll think about this a bit more soon.