Commit some initial, untested debug logging code.
This differs from what we did in Squid - here, it's a bitmap that
by definition only has meaning for each debug section that is
registered.

I aim to add a debug de-register method and a debug bitmap parse
and output methods so different modules can have a fully text
representation of their debugging flags.

But, this gets me out of the hothouse with respect to debugging;
now I can register some debug sections and start doing useful
work with it.