Start fleshing out a network buffer representation.
The eventual aim is to allow multiple underlying "things" to potentially
represent network buffers and the network IO code can choose an efficient
way to spray it out to the network.

For now comm_write() uses the network buffer directly.  Later on I'll
change it to take an iovec style array of these things so multiple
writes can be thrown together.