Add in some hack workarounds (well, kind of hack workaround) to migrate FDs to the "right" pinned thread if configured.
Use a timer to check the list every 100ms.  It sucks; instead it
should be an API call in libiapp that correctly schedules the
callback after something is scheduled.

The list handling is suboptimal - I should grab the list under
the lock and then work on thel ist, rather than doing it with
the lock held.  Again, not important for my basic tests.